tratamento de exceção em todos os métodos do Service

+++ src/main/java/br/ufpb/dcx/lab1/services/DisciplinaServices.java
@@ -1,6 +1,7 @@
 package br.ufpb.dcx.lab1.services;
 
 import br.ufpb.dcx.lab1.entities.Disciplina;
+import br.ufpb.dcx.lab1.services.exceptions.DisciplinaNotFound;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -20,22 +21,32 @@ public class DisciplinaServices {
         return d;
     }
 
-    public Disciplina findById(Integer id) throws Exception {
+    public Disciplina findById(Integer id) throws DisciplinaNotFound {
         Optional<Disciplina> d = Optional.ofNullable(dataBase.get(id));
-        return d.orElseThrow(() -> new Exception("Disciplina não encontrada!"));
+        return d.orElseThrow(() -> new DisciplinaNotFound("Não foi encontrada disciplina com esse id: "+ id));
 
     }
 
     public List<Disciplina> findAll() {
+        if (dataBase.size() == 0){
+            throw new DisciplinaNotFound("Lista de disciplinas está vazia!");
+        }
         return new ArrayList<>(dataBase.values());
     }
 
     public void delete(Integer id) {
+        Disciplina d = dataBase.get(id);
+        if (d == null){
+            throw new DisciplinaNotFound("Não foi encontrada disciplina com esse id: "+ id);
+        }
         dataBase.remove(id);
     }
 
     public Disciplina update(Disciplina obj, Integer id) {
         Disciplina novaDisciplina = dataBase.get(id);
+        if (novaDisciplina == null){
+            throw new DisciplinaNotFound("Não foi encontrada disciplina com esse id: "+ id);
+        }
         novaDisciplina.setNome(obj.getNome());
         dataBase.put(id, novaDisciplina);
         return novaDisciplina;
@@ -43,6 +54,9 @@ public class DisciplinaServices {
 
     public Disciplina addNota(Integer id, Integer nota){
         Disciplina novaDisciplina = dataBase.get(id);
+        if (novaDisciplina == null){
+            throw new DisciplinaNotFound("Não foi encontrada disciplina com esse id: "+ id);
+        }
         novaDisciplina.getNotas().add(nota);
         dataBase.put(id, novaDisciplina);
         return novaDisciplina;
@@ -50,6 +64,9 @@ public class DisciplinaServices {
 
     public  Disciplina addLike(Integer id, Integer like){
         Disciplina novaDisciplina = dataBase.get(id);
+        if (novaDisciplina == null){
+            throw new DisciplinaNotFound("Não foi encontrada disciplina com esse id: "+ id);
+        }
         int somaLike = novaDisciplina.getLikes() + 1;
         novaDisciplina.setLikes(somaLike);
         dataBase.put(id, novaDisciplina);
@@ -58,6 +75,9 @@ public class DisciplinaServices {
 
     public List<Disciplina> findRanking(){
         List<Disciplina> list = new ArrayList<>(dataBase.values());
+        if (list.size() == 0){
+            throw new DisciplinaNotFound("Lista de disciplinas está vazia!");
+        }
         Collections.sort(list,Comparator.comparingInt(Disciplina::getMedia).reversed());
         return list;
     }
